feat: Dashboard 경로 추가 및 파일 이동 - #1
